fix: changed appView.js to  appView.ts
